requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
pika==1.3.2
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
